requirements.txt
ocrmypdf==17.2.0
pdfplumber==0.11.9
pdf2image==1.17.0
pytesseract==0.3.13
pandas==3.0.1
opencv-python==4.13.0.92
beautifulsoup4==4.14.3
requests==2.32.5
selenium==4.41.0
mysql-connector-python==9.6.0
Flask==3.1.3
